requirements.txt
sqlalchemy==2.0.25
asyncio==3.4.3
asyncpg==0.29.0
python-dotenv==1.0.0
aiogram==3.3.0
